refactor(plugin-client): emit operations.ts via AST builders
Slice 001 of the core/plugin simplification: prove a generator can emit
through the @kubb/ast create* builders instead of JSX, with byte-identical
output. operationsGenerator now returns FileNode[] built with createFile/
createSource/createConst/createText and no longer imports @kubb/renderer-jsx.
The internal Operations JSX component (its only user) is removed. The
existing snapshot is unchanged.
