+++ .changeset/plugin-client-operations-builder.md
@@ -1,0 +1,5 @@
+---
+"@kubb/plugin-client": patch
+---
+
+Build the `operations.ts` aggregate file from `@kubb/ast` builders instead of a JSX component. No change to generated output.
