+++ packages/plugin-client/src/index.ts
@@ -2,7 +2,7 @@ export { Client } from './components/Client.tsx'
 export { classClientGenerator } from './generators/classClientGenerator.tsx'
 export { clientGenerator } from './generators/clientGenerator.tsx'
 export { groupedClientGenerator } from './generators/groupedClientGenerator.tsx'
-export { operationsGenerator } from './generators/operationsGenerator.tsx'
+export { operationsGenerator } from './generators/operationsGenerator.ts'
 export { staticClassClientGenerator } from './generators/staticClassClientGenerator.tsx'
 export { default, pluginClient, pluginClientName } from './plugin.ts'
 export { resolverClient } from './resolvers/resolverClient.ts'
