+++ packages/plugin-client/src/plugin.ts
@@ -7,7 +7,7 @@ import { pluginZodName } from '@kubb/plugin-zod'
 import { classClientGenerator } from './generators/classClientGenerator.tsx'
 import { clientGenerator } from './generators/clientGenerator.tsx'
 import { groupedClientGenerator } from './generators/groupedClientGenerator.tsx'
-import { operationsGenerator } from './generators/operationsGenerator.tsx'
+import { operationsGenerator } from './generators/operationsGenerator.ts'
 import { staticClassClientGenerator } from './generators/staticClassClientGenerator.tsx'
 import { resolverClient } from './resolvers/resolverClient.ts'
 import { source as axiosClientSource } from './templates/clients/axios.source.ts'
